started working on adding 32u, 64u, 64s, bool and 16bf types to OpenCV

+++ modules/core/include/opencv2/core/hal/intrin_neon.hpp
@@ -1622,7 +1622,7 @@ inline int v_signmask(const v_uint64x2& a)
     const int64x2_t signPosition = {0,1};
     uint64x2_t v0 = vshlq_u64(vshrq_n_u64(a.val, 63), signPosition);
     uint64_t t0 = vaddvq_u64(v0);
-    return t0;
+    return (int)t0;
 #else // #if CV_NEON_AARCH64
     int64x1_t m0 = vdup_n_s64(0);
     uint64x2_t v0 = vshlq_u64(vshrq_n_u64(a.val, 63), vcombine_s64(m0, m0));
